requirements.txt

python-dotenv
py-solc-x
web3==5.31.3
Flask
